+++ trionyx/trionyx/models.py
@@ -23,7 +23,7 @@ from django.template.loader import render_to_string
 from trionyx import models
 from trionyx.utils import get_current_request
 from trionyx.data import TIMEZONES
-from trionyx.trionyx import LOCAL_DATA
+from trionyx import utils
 
 
 # =============================================================================
@@ -80,7 +80,6 @@ class User(models.BaseModel, AbstractBaseUser, PermissionsMixin):
     first_name = models.CharField(_('First name'), max_length=64, blank=True, default='')
     last_name = models.CharField(_('Last name'), max_length=64, blank=True, default='')
     is_active = models.BooleanField(_('Active'), default=True)
-    is_admin = models.BooleanField(_('Is admin'), default=True)
     date_joined = models.DateTimeField(_('Date joined'), default=timezone.now)
     last_online = models.DateTimeField(_('Last online'), blank=True, null=True)
     avatar = models.ImageField(_('Avatar'), blank=True, upload_to='avatars/', default='')
@@ -154,7 +153,7 @@ class UserAttributeManager(models.Manager):
 
     def set_attribute(self, user, code, value):
         """Set attribute for user"""
-        setattr(LOCAL_DATA, 'trionyx_user_attributes', None)
+        utils.set_local_data('trionyx_user_attributes', None)
         cache.delete(f'user-attributes-{user.id}')
 
         self.update_or_create(user=user, code=code, defaults={'value': value})
@@ -163,8 +162,7 @@ class UserAttributeManager(models.Manager):
         """Get attribute for user"""
         cache_key = f'user-attributes-{user.id}'
 
-        # Only use LOCAL_DATA for request, to prevent no updates in Celery or different user attributes are used
-        attributes = getattr(LOCAL_DATA, 'trionyx_user_attributes', None) if get_current_request() else None
+        attributes = utils.get_local_data('trionyx_user_attributes')
 
         if attributes is None:
             attributes = cache.get(cache_key)
@@ -175,7 +173,7 @@ class UserAttributeManager(models.Manager):
             }
             cache.set(cache_key, attributes, timeout=60 * 60 * 24)
 
-        setattr(LOCAL_DATA, 'trionyx_user_attributes', attributes)
+        utils.set_local_data('trionyx_user_attributes', attributes)
 
         return attributes.get(code, default)